+++ src/redux/api/index.ts
@@ -1,0 +1,2 @@
+import {projectApi} from './project'
+export const apis = [projectApi]